+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [1.1.1] - 2024-08-11
+
 ### Changed
 
 - Fix commit checkin handler for multi-module projects
@@ -64,7 +66,8 @@
 - The ability to apply Spotless formatting to the currently opened file in the editor or to the entire project.
 - Introduced an option to execute the default "optimize import" task before applying Spotless formatting.
 
-[Unreleased]: https://github.com/lipiridi/spotless-applier/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/lipiridi/spotless-applier/compare/v1.1.1...HEAD
+[1.1.1]: https://github.com/lipiridi/spotless-applier/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/lipiridi/spotless-applier/compare/v1.0.5...v1.1.0
 [1.0.5]: https://github.com/lipiridi/spotless-applier/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/lipiridi/spotless-applier/compare/v1.0.3...v1.0.4